pyproject.toml
[project]
name = "veille-parlementaire-sport"
version = "0.1.0"
description = "Veille parlementaire et institutionnelle automatisée — secteur sport (Sideline Conseil)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "feedparser>=6.0",
  "lxml>=5.2",
  "beautifulsoup4>=4.12",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "tenacity>=9.0",
  "jinja2>=3.1",
  "python-dateutil>=2.9",
  "unidecode>=1.3",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
veille = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 110
target-version = "py311"
